+++ .gitignore
@@ -1,0 +1,3 @@
+java-test/.idea
+java-test/target
+java-test/java-test.iml
